Fix bug in C API for bool parameters. Fix bug on irish-electricty (#25)
Since C does not have a boolean type, we need to allow `cuOptGetIntParameter` and `cuOptSetIntParameter` to work for boolean parameters as well.

On irish-electricity we had a bug where we were stopping PDLP early, due to a numerical error inside dual simplex. In concurrent mode we should only stop the other solver when we have reached a definitive conclusion on the problem.

Fix a bug in the Python docstrings found by Flora.

Also log whenever we change settings. And set the log pattern to `%v` outside the solve as well.

Authors:
  - Chris Maes (https://github.com/chris-maes)

Approvers:
  - Trevor McKay (https://github.com/tmckayus)
  - Rajesh Gandham (https://github.com/rg20)

URL: https://github.com/NVIDIA/cuopt/pull/25