Bug fix: solution incorrect when simplex presolve removes empty columns (#29)
Fixes an issue reported on the following MPS file, where `_dummy` variable does not appear in the constraints. We remove empty columns (and thus `_dummy`) in presolve for dual simplex. But do not take this into account when constructing the solution. 

Further work, is needed to develop an actual LP presolve. 

```
*SENSE:Minimize
NAME          test_export_dict_LP_no_obj
ROWS
 N  OBJ
 G  c1
 E  c2
 L  c3
 G  c4
COLUMNS
    __dummy   OBJ        1.000000000000e+00
    w         c4         1.000000000000e+00
    x         c1         1.000000000000e+00
    x         c2         1.000000000000e+00
    y         c1         1.000000000000e+00
    y         c3        -1.000000000000e+00
    z         c2         1.000000000000e+00
    z         c3         1.000000000000e+00
RHS
    RHS       c1         5.000000000000e+00
    RHS       c2         1.000000000000e+01
    RHS       c3         7.000000000000e+00
    RHS       c4         0.000000000000e+00
BOUNDS
 FX BND       __dummy    0.000000000000e+00
 FX BND       w          0.000000000000e+00
 UP BND       x          4.000000000000e+00
 LO BND       y         -1.000000000000e+00
 UP BND       y          1.000000000000e+00
```

Fixes #13  which was caused by an issue in column scaling when the norm of a column is zero.

Authors:
  - Chris Maes (https://github.com/chris-maes)

Approvers:
  - Rajesh Gandham (https://github.com/rg20)
  - Nicolas Blin (https://github.com/Kh4ster)

URL: https://github.com/NVIDIA/cuopt/pull/29