Disable failing FJ test on ARM CI (#39)
Temporary band-aid for https://github.com/rapidsai/cuopt/issues/2489, as the underlying issue is difficult to debug and possibly a CUB/cuSparse bug

Under high workloads, it appears calls to cusparse's CSR transpose algorithm return an off-by-one error in one of the indices. This only manifests in the ARM CI in the FJ tests on a specific instance. This PR disables the tests on this instance when building on ARM to work around this issue until the root bug is fixed.

Authors:
  - Alice Boucher (https://github.com/aliceb-nv)

Approvers:
  - Rajesh Gandham (https://github.com/rg20)

URL: https://github.com/NVIDIA/cuopt/pull/39