Fix issue with reported solving time on time-limit termination (#93)
Solving statistics information was not passed along to the
mip_solution_t object when terminating on a timeout, causing
cuOptGetSolveTime to incorrectly return the default value of 0.
This PR also fixes a conversion issue with `get_solve_time()` being
treated as milliseconds instead of seconds when running in LP mode.

Closes #91